+++ History.md
@@ -3,6 +3,14 @@ bunyan-logzio changes
 
 Documentation of all changes to bunyan-logzio.
 
+v0.0.2
+------
+
+- removed `buffer` option
+  - not needed since logzio buffers by default
+- added timeout to tests so there is enough time to flush logs to logzio
+
+
 v0.0.1
 ------